重构 Operator
为运算符增加优先级的概念。这样在做运算法聚合的时候，大部分运算法都是二元符，只需要按照优先级从高到低不断聚合。三元运算符的优先级最低，并且是三元的，进行下特殊处理。
这样token聚合的逻辑就变得清晰而且简单。
不仅如此，引入了操作符优先级之后，计算更有了正确性保证。之前的代码，对于比较运算符等，都没有正确的处理，全部都是左优先完成。但是实际上如果遇到一些加减符号的时候，是要优先处理加减符号的。所以引入这个优先级的方式处理起来更容易了。代码也得到了统一。

+++ src/test/java/com/jfirer/jfireel/BracketTest.java
@@ -1,25 +1,27 @@
 package com.jfirer.jfireel;
 
-import static org.junit.Assert.assertEquals;
+import com.jfirer.jfireel.expression.Expression;
+import org.junit.Test;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.junit.Test;
-import com.jfirer.jfireel.expression.Expression;
+
+import static org.junit.Assert.assertEquals;
 
 public class BracketTest
 {
     @Test
     public void test()
     {
-        int[] array = new int[] { 1, 2, 3, 4 };
-        Map<String, Object> vars = new HashMap<String, Object>();
+        int[]               array = new int[]{1, 2, 3, 4};
+        Map<String, Object> vars  = new HashMap<String, Object>();
         vars.put("array", array);
         Expression lexer = Expression.parse("array[2]");
-        assertEquals(3, lexer.calculate(vars));
+        assertEquals(3, ((Integer) lexer.calculate(vars)).intValue());
     }
-    
+
     @Test
     public void test2()
     {
@@ -33,7 +35,7 @@ public class BracketTest
         Expression lexer2 = Expression.parse("map[age]");
         assertEquals("12", lexer2.calculate(vars));
     }
-    
+
     @Test
     public void test3()
     {
@@ -45,7 +47,7 @@ public class BracketTest
         Expression lexer = Expression.parse("list[1]");
         assertEquals("13", lexer.calculate(vars));
     }
-    
+
     @Test
     public void test4()
     {
